+++ src/components/Pagination/Pagination.tsx
@@ -27,7 +27,7 @@ type Props = {
   totalCount: number;
 };
 
-export const Pagination = ({ currentPage, totalCount }: Props) => {
+export default function Pagination({ currentPage, totalCount }: Props) {
   const router = useRouter();
 
   const handlePageChange = (page: number) => {
@@ -47,7 +47,7 @@ export const Pagination = ({ currentPage, totalCount }: Props) => {
   };
 
   return <StyledRcPagination itemRender={itemRender} total={totalCount} />;
-};
+}
 
 export const StyledRcPagination = styled(RcPagination)`
   display: flex;
