+++ src/pages/index.tsx
@@ -2,10 +2,10 @@ import { GetStaticProps } from 'next';
 import Head from 'next/head';
 import styled from 'styled-components';
 import Card from '@/components/Card';
+import Pagination from '@/components/Pagination';
 import { microcms } from '@/libs/microcms';
 import { Blog, Post } from '@/types/blog';
 import MaxWidthWrapper from 'src/components/MaxWidthWrapper';
-import { Pagination } from 'src/components/home/Pagination';
 
 type Props = {
   blog: Post[];
